+++ src/components/EditSections/EditReadingRoomAccess/constants.js
@@ -1,0 +1,17 @@
+export const rraColumns = {
+  ACCESS: 'access',
+  READING_ROOM_NAME: 'readingRoomName',
+  NOTES: 'notes',
+};
+
+// TODO: Translations to be applied to the options
+export const READING_ROOM_ACCESS_OPTIONS = [
+  {
+    label: 'Allowed',
+    value: 'ALLOWED'
+  },
+  {
+    label: 'Not allowed',
+    value: 'NOT_ALLOWED'
+  }
+];